move aws interface to separate file
